+++ SemiconductorTestLibrary.Extensions/source/InstrumentAbstraction/Scope/Control.cs
@@ -1,0 +1,61 @@
+using static NationalInstruments.SemiconductorTestLibrary.Common.ParallelExecution;
+
+namespace NationalInstruments.SemiconductorTestLibrary.InstrumentAbstraction.Scope
+{
+    /// <summary>
+    /// Defines methods for performing measurements on the NI-Scope session.
+    /// </summary>
+    public static class Control
+    {
+        #region methods on ScopeSessionsBundle
+
+        /// <summary>
+        /// Aborts an in-progress acquisition on all sessions in the bundle.
+        /// </summary>
+        /// <param name="sessionsBundle">The <see cref="ScopeSessionsBundle"/> object.</param>
+        public static void Abort(this ScopeSessionsBundle sessionsBundle)
+        {
+            sessionsBundle.Do(sessionInfo =>
+            {
+                sessionInfo.Session.Measurement.Abort();
+            });
+        }
+
+        /// <summary>
+        /// Performs auto-setup on all sessions in the bundle.
+        /// </summary>
+        /// <param name="sessionsBundle">The <see cref="ScopeSessionsBundle"/> object.</param>
+        public static void AutoSetup(this ScopeSessionsBundle sessionsBundle)
+        {
+            sessionsBundle.Do(sessionInfo =>
+            {
+                sessionInfo.Session.Measurement.AutoSetup();
+            });
+        }
+
+        /// <summary>
+        /// Commits the acquisition on all sessions in the bundle.
+        /// </summary>
+        /// <param name="sessionsBundle">The <see cref="ScopeSessionsBundle"/> object.</param>
+        public static void Commit(this ScopeSessionsBundle sessionsBundle)
+        {
+            sessionsBundle.Do(sessionInfo =>
+            {
+                sessionInfo.Session.Measurement.Commit();
+            });
+        }
+
+        /// <summary>
+        /// Starts acquisition on all sessions in the bundle.
+        /// </summary>
+        /// <param name="sessionsBundle">The <see cref="ScopeSessionsBundle"/> object.</param>
+        public static void Initiate(this ScopeSessionsBundle sessionsBundle)
+        {
+            sessionsBundle.Do(sessionInfo =>
+            {
+                sessionInfo.Session.Measurement.Initiate();
+            });
+        }
+        #endregion methods on ScopeSessionsBundle
+    }
+}
